backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
pydantic-settings
supabase
finnhub-python
yfinance>=0.2
pandas
numpy
scipy
scikit-learn
python-jose[cryptography]
passlib[bcrypt]
httpx
python-multipart
cachetools
python-dotenv
requests
